waste api/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.2.0
python-multipart>=0.0.9
numpy>=1.26.3
python-dotenv>=1.0.0
httpx>=0.25.0 